+++ AromasWeb/AromasWeb.AccesoADatos/Modelos/RecetaInsumo.cs
@@ -8,8 +8,8 @@ namespace AromasWeb.AccesoADatos.Modelos
         public int IdInsumo { get; set; }
         public int IdReceta { get; set; }
         public decimal CantidadUtilizada { get; set; }
-        public decimal CostoUnitario { get; set; }
-        public decimal CostoTotalIngrediente { get; set; }
+        public int CostoUnitario { get; set; }
+        public int CostoTotalIngrediente { get; set; }
 
         // Propiedades de navegación
         public virtual RecetaAD Receta { get; set; }
